+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Added an option to disable currency conversion.
 - Fixed a bug where popups were not scrollable when no tabs are shown.
 - Force Google Docs to use HTML mode for now anyway.
+- Ignore zero-width non-joiner characters (which Google Docs sometimes likes to
+  put between characters)
 
 ## 1.4.2 (2021-09-24)